+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-node.js
@@ -15,7 +15,7 @@ jest.mock(`fs`, () => {
 jest.mock(`sharp`, () => {
   let sharp = jest.fn(
     () =>
-      new class {
+      new (class {
         resize() {
           return this
         }
@@ -28,7 +28,7 @@ jest.mock(`sharp`, () => {
             height: 128,
           }
         }
-      }()
+      })()
   )
 
   sharp.simd = jest.fn()