+++ packages/gatsby-source-wordpress/src/normalize.js
@@ -479,6 +479,7 @@ exports.downloadMediaFiles = async ({
   touchNode,
   getNode,
   _auth,
+  reporter,
 }) =>
   Promise.all(
     entities.map(async e => {
@@ -513,6 +514,7 @@ exports.downloadMediaFiles = async ({
               createNodeId,
               parentNodeId: e.id,
               auth: _auth,
+              reporter,
             })
 
             if (fileNode) {