+++ packages/gatsby/src/query/__tests__/file-parser.js
@@ -1,12 +1,13 @@
 jest.mock(`fs-extra`, () => {
-  const fs = jest.requireActual(`fs`)
+  const fs = jest.requireActual(`fs-extra`)
   return {
     ...fs,
     readFile: jest.fn(),
   }
 })
 jest.mock(`../../utils/api-runner-node`, () => () => [])
-
+jest.mock(`gatsby-cli/lib/reporter/index`)
+const reporter = require(`gatsby-cli/lib/reporter`)
 const fs = require(`fs-extra`)
 
 const FileParser = require(`../file-parser`).default
@@ -142,6 +143,45 @@ export default () => (
     render={data => <div>{data.pizza}</div>}
   />
 )`,
+    "static-query-hooks.js": `import { graphql, useStaticQuery } from 'gatsby'
+export default () => {
+  const data = useStaticQuery(graphql\`query StaticQueryName { foo }\`);
+  return <div>{data.doo}</div>;
+}`,
+    "static-query-hooks-with-type-parameter.ts": `import { graphql, useStaticQuery } from 'gatsby'
+export default () => {
+  const data = useStaticQuery<HomepageQuery>(graphql\`query StaticQueryName { foo }\`);
+  return <div>{data.doo}</div>;
+}`,
+    "static-query-hooks-missing-argument.js": `import { graphql, useStaticQuery } from 'gatsby'
+export default () => {
+  const data = useStaticQuery();
+  return <div>{data.doo}</div>;
+}`,
+    "static-query-hooks-in-separate-variable.js": `import React from "react"
+import { useStaticQuery, graphql } from "gatsby"
+
+const query = graphql\`{ allMarkdownRemark { blah { node { cheese }}}}\`
+
+export default () => {
+  const data = useStaticQuery(query);
+  return <div>{data.doo}</div>;
+}`,
+    "static-query-hooks-not-defined.js": `import React from "react"
+import { useStaticQuery, graphql } from "gatsby"
+
+export default () => {
+  const data = useStaticQuery(strangeQueryName);
+  return <div>{data.pizza}</div>;
+}`,
+    "static-query-hooks-imported.js": `import React from "react"
+import { useStaticQuery, graphql } from "gatsby"
+import strangeQueryName from "./another-file.js"
+
+export default () => {
+  const data = useStaticQuery(strangeQueryName);
+  return <div>{data.pizza}</div>;
+}`,
   }
 
   const parser = new FileParser()
@@ -153,16 +193,8 @@ export default () => (
   })
 
   it(`extracts query AST correctly from files`, async () => {
-    const spyStderr = jest.spyOn(process.stderr, `write`)
     const results = await parser.parseFiles(Object.keys(MOCK_FILE_INFO))
     expect(results).toMatchSnapshot()
-    expect(
-      spyStderr.mock.calls
-        .filter(c => c[0].includes(`warning`))
-        // Remove console colors + trim whitespace
-        // eslint-disable-next-line
-        .map(c => c[0].replace(/\x1B[[(?);]{0,2}(;?\d)*./g, ``).trim())
-    ).toMatchSnapshot()
-    spyStderr.mockRestore()
+    expect(reporter.warn).toMatchSnapshot()
   })
 })