+++ www/src/utils/typography.js
@@ -2,51 +2,65 @@ import Typography from "typography"
 import CodePlugin from "typography-plugin-code"
 import {
   space,
-  fontSizes,
+  fontSizes as fontSizeTokens,
   colors,
   transition,
   radii,
-  breakpoints,
+  mediaQueries,
   lineHeights,
   letterSpacings,
   fonts,
+  fontWeights,
 } from "./tokens"
+
+const fontSizes = fontSizeTokens.map(token => `${token / 16}rem`)
 
 const _options = {
   bodyFontFamily: fonts.system,
   headerFontFamily: fonts.header,
-  monospaceFontFamily: fonts.monospace.join(`,`),
-  systemFontFamily: fonts.system,
   baseLineHeight: lineHeights.default,
   headerLineHeight: lineHeights.dense,
-  headerColor: colors.gray.dark,
-  bodyColor: colors.gray.copy,
+  headerColor: colors.text.header,
+  bodyColor: colors.text.primary,
   plugins: [new CodePlugin()],
   overrideStyles: ({ rhythm }) => {
     return {
+      html: {
+        backgroundColor: colors.white,
+        WebkitFontSmoothing: `antialiased`,
+        MozOsxFontSmoothing: `grayscale`,
+      },
       a: {
         textDecoration: `none`,
+      },
+      h1: {
+        fontWeight: fontWeights[2],
       },
       "h1, h2, h3, h4, h5, h6": {
         letterSpacing: letterSpacings.tight,
       },
-      h1: { color: `#000` },
       h2: {
         marginTop: rhythm(space[9]),
       },
       h3: {
         marginTop: rhythm(space[9]),
       },
+      "h4, h5, h6": { fontSize: fontSizes[3] },
+      "h5, h6": { fontWeight: fontWeights[0] },
+      h6: { fontSize: fontSizes[2] },
       blockquote: {
         paddingLeft: rhythm(space[6]),
         marginLeft: 0,
-        borderLeft: `${rhythm(space[1])} solid ${colors.ui.light}`,
+        borderLeft: `${rhythm(space[1])} solid ${colors.ui.border.subtle}`,
       },
       hr: {
-        backgroundColor: colors.ui.light,
+        backgroundColor: colors.ui.border.subtle,
       },
       iframe: {
         border: 0,
+      },
+      "th, td": {
+        borderColor: colors.ui.border.subtle,
       },
       "tt, code, kbd, samp": {
         // reset line-height set by
@@ -54,7 +68,7 @@ const _options = {
         lineHeight: `inherit`,
       },
       "h1 code, h2 code, h3 code, h4 code, h5 code, h6 code": {
-        fontWeight: `normal`,
+        fontWeight: fontWeights[0],
         fontSize: `82.5%`,
       },
       "tt, code, kbd": {
@@ -73,7 +87,6 @@ const _options = {
       // gatsby-remark-prismjs styles
       ".gatsby-highlight": {
         background: colors.code.bg,
-        borderRadius: `${radii[1]}px`,
         position: `relative`,
         WebkitOverflowScrolling: `touch`,
       },
@@ -86,7 +99,7 @@ const _options = {
       ".gatsby-highlight pre[class*='language-']::before": {
         background: `#ddd`,
         borderRadius: `0 0 ${radii[2]}px ${radii[2]}px`,
-        color: colors.gray.dark,
+        color: colors.text.header,
         fontSize: fontSizes[0],
         fontFamily: fonts.monospace.join(`,`),
         letterSpacing: letterSpacings.tracked,
@@ -114,17 +127,20 @@ const _options = {
         content: `'GraphQL'`,
         background: `#E10098`,
         color: colors.white,
-        fontWeight: `400`,
       },
       ".gatsby-highlight pre[class='language-html']::before": {
         content: `'html'`,
         background: `#005A9C`,
         color: colors.white,
-        fontWeight: `400`,
       },
       ".gatsby-highlight pre[class='language-css']::before": {
         content: `'css'`,
         background: `#ff9800`,
+        color: colors.white,
+      },
+      ".gatsby-highlight pre[class='language-mdx']::before": {
+        content: `'mdx'`,
+        background: `#f9ac00`,
         color: colors.white,
         fontWeight: `400`,
       },
@@ -189,7 +205,6 @@ const _options = {
       },
       ".gatsby-highlight pre::-webkit-scrollbar-track": {
         background: colors.code.border,
-        borderRadius: `0 0 ${radii[2]}px ${radii[2]}px`,
       },
       // Target image captions.
       // This is kind of a fragile selector...
@@ -200,40 +215,45 @@ const _options = {
         marginBottom: rhythm(space[9]),
         display: `block`,
         fontStyle: `normal`,
-        color: colors.gray.calm,
+        color: colors.text.secondary,
         position: `relative`,
       },
       ".gatsby-resp-image-link + em a, .gatsby-resp-image-wrapper + em a": {
-        fontWeight: `normal`,
+        fontWeight: fontWeights[0],
         color: colors.lilac,
       },
       ".main-body a": {
-        color: colors.lilac,
+        color: colors.link.color,
         textDecoration: `none`,
         transition: `all ${transition.speed.fast} ${transition.curve.default}`,
-        borderBottom: `1px solid ${colors.lilac}`,
+        borderBottom: `1px solid ${colors.link.border}`,
       },
       ".main-body a:hover": {
-        borderBottomColor: colors.ui.bright,
+        borderBottomColor: colors.link.hoverBorder,
+      },
+      ".post-body h1": {
+        fontWeight: fontWeights[1],
       },
       ".post-body figure img": {
         marginBottom: 0,
       },
       ".post-body figcaption": {
-        color: colors.gray.calm,
+        color: colors.text.secondary,
         fontSize: `87.5%`,
         marginTop: rhythm(space[1]),
         marginBottom: rhythm(space[3]),
       },
+      //
       ".main-body a.anchor": {
         color: `inherit`,
-        fill: colors.lilac,
+        fill: colors.link.color,
         textDecoration: `none`,
         borderBottom: `none`,
       },
       ".main-body a.anchor:hover": {
         background: `none`,
       },
+      // gatsby-image
       ".main-body a.gatsby-resp-image-link": {
         borderBottom: `transparent`,
         marginTop: rhythm(space[9]),
@@ -258,20 +278,20 @@ const _options = {
         background: colors.code.bg,
         borderBottom: `1px solid ${colors.code.border}`,
         color: colors.code.text,
-        padding: `${rhythm(space[6])} ${rhythm(space[6])} ${rhythm(space[3])}`,
-        fontSize: `74%`,
+        padding: `${rhythm(space[5])} ${rhythm(space[6])} ${rhythm(space[4])}`,
+        fontSize: fontSizes[0],
       },
-      "@media (max-width:634px)": {
-        ".gatsby-highlight, .gatsby-code-title, .gatsby-resp-image-link": {
-          borderRadius: 0,
-          borderLeft: 0,
-          borderRight: 0,
-        },
-      },
-      [`${breakpoints.md} and (max-width:980px)`]: {
-        ".gatsby-highlight, .gatsby-code-title": {
+      [mediaQueries.md]: {
+        ".gatsby-highlight, .gatsby-resp-image-link, .gatsby-code-title": {
           marginLeft: 0,
           marginRight: 0,
+          borderRadius: `${radii[2]}px`,
+        },
+        ".gatsby-code-title": {
+          borderRadius: `${radii[2]}px ${radii[2]}px 0 0`,
+        },
+        ".gatsby-code-title + .gatsby-highlight": {
+          borderRadius: `0 0 ${radii[2]}px ${radii[2]}px`,
         },
       },
       video: {
@@ -284,37 +304,13 @@ const _options = {
       ".egghead-video": {
         border: `none`,
       },
-      [breakpoints.lg]: {
+      [mediaQueries.lg]: {
         ".gatsby-highlight, .post-body .gatsby-resp-image-link, .gatsby-code-title": {
-          marginLeft: rhythm(-space[7]),
-          marginRight: rhythm(-space[7]),
-        },
-        ".gatsby-highlight pre": {
-          padding: `${rhythm(space[7])} 0`,
-          marginBottom: rhythm(space[7]),
-        },
-        ".gatsby-highlight pre code": {
-          padding: `0 ${rhythm(space[7])}`,
-        },
-        ".gatsby-highlight-code-line": {
-          marginRight: rhythm(-space[7]),
-          marginLeft: rhythm(-space[7]),
-          paddingRight: rhythm(space[7]),
-          paddingLeft: rhythm(space[6]),
-          borderLeftWidth: rhythm(space[2]),
-        },
-        ".gatsby-code-title": {
-          marginRight: rhythm(-space[7]),
-          marginLeft: rhythm(-space[7]),
-          padding: `${rhythm(space[6])} ${rhythm(space[7])} ${rhythm(
-            space[3]
-          )}`,
-        },
-        ".gatsby-highlight pre[class*='language-']::before": {
-          right: rhythm(space[7]),
+          marginLeft: rhythm(-space[6]),
+          marginRight: rhythm(-space[6]),
         },
       },
-      [breakpoints.xxl]: {
+      [mediaQueries.xxl]: {
         html: {
           fontSize: `${(18 / 16) * 100}%`,
         },
@@ -350,10 +346,10 @@ const _options = {
         color: colors.code.cssString,
       },
       ".token.important": {
-        fontWeight: `normal`,
+        fontWeight: fontWeights[0],
       },
       ".token.bold": {
-        fontWeight: `bold`,
+        fontWeight: fontWeights[1],
       },
       ".token.italic": {
         fontStyle: `italic`,